Add quick slot gui, AuthenticPlayer can use quick slot by number keys, minor changes applied

+++ evergreen_client/PlayerStatusGUI.cpp
@@ -18,7 +18,7 @@ PlayerStatusGUI::PlayerStatusGUI(const std::shared_ptr<udsdx::SceneObject>& obje
         m_healthFillRenderer = m_healthFill->AddComponent<GUIImage>();
         m_healthFill->GetTransform()->SetLocalPosition(Vector3(-622.5f, -480.0f, 0.0f));
         m_healthFillRenderer->SetTexture(INSTANCE(Resource)->Load<udsdx::Texture>(RESOURCE_PATH(L"gui\\health_fill.png")), true);
-        m_healthFillWidthCache = m_healthFillRenderer->GetSize().x;
+        m_healthFillWidthCache = static_cast<float>(m_healthFillRenderer->GetSize().x);
     }
 
     {
@@ -46,7 +46,7 @@ void PlayerStatusGUI::SetCurrentHealth(int value)
 	m_currentHealth = value;
     float factor = std::clamp(static_cast<float>(m_currentHealth) / static_cast<float>(m_maxHealth), 0.0f, 1.0f);
     Vector2Int size = m_healthFillRenderer->GetSize();
-    size.x = m_healthFillWidthCache * factor;
+    size.x = static_cast<int32_t>(m_healthFillWidthCache * factor);
     m_healthFillRenderer->SetSize(size);
     m_healthFill->GetTransform()->SetLocalPosition(Vector3(-622.5f + (size.x - m_healthFillWidthCache) / 2.0f, -480.0f, 0.0f));
     m_textRenderer->SetText(std::format(L"{0:02} / {1:02}", m_currentHealth, m_maxHealth));
